Add required `combine_fun` param to `ForBindings`.
Along the way, we also make it and `UniversalTransform` generic. Maybe
all of the universal transforms should be generic?

By requiring the user to provide a `combine_fun` we ensure that the
result of application should always have type `DataResult` as expected.

`GroupByBindings` is now just a type-specialized wrapper around
`ForBindings` with a `DataFrame`-specific `combine_fun`.