Deploy current docs: Force separate cache keys for linters/docs.
mypy and pytest will use the same cache key.

It matters because github won't overwrite the cache on a cache hit, so
if linters runs first (usually does), pytest etc will be stuck using a
pip cache with a much smaller set of packages in it. ffd22b7705987e350317097d232f9b98c8238279